+++ src/components/main/main.jsx
@@ -2,8 +2,8 @@ import React, {useEffect} from 'react';
 import {Link, useHistory} from 'react-router-dom';
 import {useDispatch, useSelector} from 'react-redux';
 import {getFilteredMovies} from '../../store/selectors/selectors';
-import {AuthorizationStatus, AVATAR_URL} from '../../const';
-import {AppRoute, getPlayerUrl} from '../../routes';
+import {AuthorizationStatus, AVATAR_URL, AppRoute} from '../../util/const';
+import {getPlayerUrl} from '../../util/route';
 import MovieList from '../movie-list/movie-list';
 import GenreList from '../genre-list/genre-list';
 import Loading from '../loading/loading';
