+++ src/components/not-found/not-found.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
-import {AppRoute} from '../../routes';
+import {AppRoute} from '../../util/const';
 
 const NotFound = () => {
   return (
