+++ src/components/private-route/private-route.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import {Redirect, Route} from 'react-router-dom';
 import {useSelector} from 'react-redux';
 import {PropTypes} from 'prop-types';
-import {AuthorizationStatus} from '../../const';
-import {AppRoute} from '../../routes';
+import {AuthorizationStatus} from '../../util/const';
+import {AppRoute} from '../../util/const';
 
 const PrivateRoute = ({render, path, exact}) => {
   const {authorizationStatus} = useSelector((state) => state.USER);
