--- src/const.js
@@ -1,8 +1,0 @@
-export const AVATAR_URL = `https://i.pravatar.cc/63`;
-
-export const FILTER_ALL_GENRES = `All genres`;
-
-export const AuthorizationStatus = {
-  AUTH: `AUTH`,
-  NO_AUTH: `NO_AUTH`
-};
