+++ src/util/browser-history.js
