+++ src/util/const.js
@@ -1,0 +1,21 @@
+export const AVATAR_URL = `https://i.pravatar.cc/63`;
+
+export const FILTER_ALL_GENRES = `All genres`;
+
+export const AuthorizationStatus = {
+  AUTH: `AUTH`,
+  NO_AUTH: `NO_AUTH`
+};
+
+export const APIRoute = {
+  ROOT: `/`,
+  LOGIN: `/login`,
+  MOVIES: `/films`
+};
+
+export const AppRoute = {
+  ROOT: `/`,
+  LOGIN: `/login`,
+  MYLIST: `/mylist`,
+  MOVIES: `/films`,
+};
